refactor: move restful api utils to common level
